pyproject.toml
[project]
name = "api-demo"
version = "0.2.1"
dependencies = [
  'requests="2.31.0"',
  "click",
  "python-dotenv"
]
requires-python = ">= 3.10"

[project.scripts]
bird = "main:lookup_num"
album = "main:get_photos"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"
